refactor: canister-details folder name
