+++ frontend/svelte/src/lib/components/canister-detail/ControllersCard.svelte
