+++ frontend/svelte/src/lib/components/canister-detail/CyclesCard.svelte
