+++ frontend/svelte/src/lib/components/canister-detail/DetachCanisterButton.svelte
