+++ frontend/svelte/src/lib/components/canister-detail/NewControllerReview.svelte
