+++ frontend/svelte/src/lib/components/canister-detail/RemoveCanisterControllerButton.svelte
