+++ frontend/svelte/src/routes/CanisterDetail.svelte
@@ -18,8 +18,8 @@
   import { replacePlaceholders, translate } from "../lib/utils/i18n.utils";
   import SkeletonParagraph from "../lib/components/ui/SkeletonParagraph.svelte";
   import SkeletonCard from "../lib/components/ui/SkeletonCard.svelte";
-  import CyclesCard from "../lib/components/canister_details/CyclesCard.svelte";
-  import ControllersCard from "../lib/components/canister_details/ControllersCard.svelte";
+  import CyclesCard from "../lib/components/canister-detail/CyclesCard.svelte";
+  import ControllersCard from "../lib/components/canister-detail/ControllersCard.svelte";
   import SkeletonTitle from "../lib/components/ui/SkeletonTitle.svelte";
   import { writable } from "svelte/store";
   import {
@@ -31,7 +31,7 @@
   import type { CanisterDetails } from "../lib/canisters/ic-management/ic-management.canister.types";
   import AddCyclesModal from "../lib/modals/canisters/AddCyclesModal.svelte";
   import Toolbar from "../lib/components/ui/Toolbar.svelte";
-  import DetachCanisterButton from "../lib/components/canister_details/DetachCanisterButton.svelte";
+  import DetachCanisterButton from "../lib/components/canister-detail/DetachCanisterButton.svelte";
   import { toastsStore } from "../lib/stores/toasts.store";
   import { busy } from "../lib/stores/busy.store";
   import { getCanisterFromStore } from "../lib/utils/canisters.utils";
