+++ frontend/svelte/src/tests/lib/components/canister_details/CyclesCard.spec.ts
@@ -3,7 +3,7 @@
  */
 
 import { render } from "@testing-library/svelte";
-import CyclesCard from "../../../../lib/components/canister_details/CyclesCard.svelte";
+import CyclesCard from "../../../../lib/components/canister-detail/CyclesCard.svelte";
 import { formatCyclesToTCycles } from "../../../../lib/utils/canisters.utils";
 import en from "../../../mocks/i18n.mock";
